Add a ruler to workspace vscode settings
An exception to PEP 8 is our rules on line lengths.
Don’t limit lines of code to 79 characters if it means the code looks
significantly uglier or is harder to read.
We allow up to 119 characters as this is the width of
GitHub code review; anything longer requires horizontal
scrolling which makes review more difficult.

This check is included when you run flake8.
Documentation, comments, and docstrings should be wrapped at 79
characters, even though PEP 8 suggests 72.
Ref: https://docs.djangoproject.com/en/dev/internals/contributing/writing-code/coding-style/